rtoshat

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -23,8 +23,8 @@ return new class extends Migration
             $table->integer('length')->nullable();
             $table->integer('age')->nullable();
             $table->double('BMI')->nullable();
-            $table->double('total_time_practice')->nullable();
-            $table->double('total_calorie')->nullable();
+            $table->integer('total_time_practice')->nullable();
+            $table->integer('total_calorie')->nullable();
             $table->string('image')->nullable();
             $table->integer('points')->nullable();
             //$table->foreignId('plan_id')->constrained('plans')->cascadeOnDelete();
